broke out the preferences and config
