+++ pkg/application/application_test.go
@@ -77,7 +77,7 @@ func TestGetDefaultAppName(t *testing.T) {
 			r, _ := regexp.Compile(tt.wantRE)
 			match := r.MatchString(name)
 			if !match {
-				fetchedConfig, _ := config.New()
+				fetchedConfig, _ := preference.New()
 				t.Errorf("randomly generated application name %s does not match regexp %s and config is %+v\nthe prefix is %s", name, tt.wantRE, fetchedConfig, *fetchedConfig.OdoSettings.NamePrefix)
 			}
 		})
