+++ pkg/component/component.go
@@ -15,9 +15,9 @@ import (
 	"github.com/pkg/errors"
 	applabels "github.com/redhat-developer/odo/pkg/application/labels"
 	componentlabels "github.com/redhat-developer/odo/pkg/component/labels"
-	"github.com/redhat-developer/odo/pkg/config"
 	"github.com/redhat-developer/odo/pkg/log"
 	"github.com/redhat-developer/odo/pkg/occlient"
+	"github.com/redhat-developer/odo/pkg/preference"
 	"github.com/redhat-developer/odo/pkg/storage"
 	urlpkg "github.com/redhat-developer/odo/pkg/url"
 	"github.com/redhat-developer/odo/pkg/util"
@@ -73,7 +73,7 @@ func GetDefaultComponentName(componentPath string, componentPathType occlient.Cr
 	}
 
 	// Fetch config
-	cfg, err := config.New()
+	cfg, err := preference.New()
 	if err != nil {
 		return "", errors.Wrap(err, "unable to generate random component name")
 	}
@@ -275,7 +275,7 @@ func Delete(client *occlient.Client, componentName string, applicationName strin
 	s := log.Spinnerf("Deleting component %s", componentName)
 	defer s.End(false)
 
-	cfg, err := config.New()
+	cfg, err := preference.New()
 	if err != nil {
 		return errors.Wrapf(err, "unable to create new configuration to delete %s", componentName)
 	}
@@ -327,7 +327,7 @@ func Delete(client *occlient.Client, componentName string, applicationName strin
 
 // SetCurrent sets the given componentName as active component
 func SetCurrent(componentName string, applicationName string, projectName string) error {
-	cfg, err := config.New()
+	cfg, err := preference.New()
 	if err != nil {
 		return errors.Wrapf(err, "unable to set current component %s", componentName)
 	}
@@ -343,7 +343,7 @@ func SetCurrent(componentName string, applicationName string, projectName string
 // GetCurrent component in active application
 // returns "" if there is no active component
 func GetCurrent(applicationName string, projectName string) (string, error) {
-	cfg, err := config.New()
+	cfg, err := preference.New()
 	if err != nil {
 		return "", errors.Wrap(err, "unable to get config")
 	}
