+++ pkg/odo/cli/config/config.go
@@ -5,6 +5,8 @@ import (
 
 	"github.com/redhat-developer/odo/pkg/config"
 	"github.com/redhat-developer/odo/pkg/odo/util"
+	"github.com/redhat-developer/odo/pkg/preference"
+
 	"github.com/spf13/cobra"
 	ktemplates "k8s.io/kubernetes/pkg/kubectl/cmd/templates"
 )
@@ -26,7 +28,7 @@ func NewCmdConfiguration(name, fullName string) *cobra.Command {
 	configurationCmd := &cobra.Command{
 		Use:   name,
 		Short: "Modifies configuration settings",
-		Long:  fmt.Sprintf(configLongDesc, config.FormatSupportedParameters(), config.FormatLocallySupportedParameters()),
+		Long:  fmt.Sprintf(configLongDesc, preference.FormatSupportedParameters(), config.FormatLocallySupportedParameters()),
 		Example: fmt.Sprintf("%s\n%s\n%s",
 			configurationViewCmd.Example,
 			configurationSetCmd.Example,
